+++ experimental/builder/include/ck_tile/builder/conv_algorithm_limits.hpp
@@ -15,11 +15,16 @@ concept InputVectorTransferLimits = requires {
                      Value.lds_dst_scalar_per_vector > 0;
 };
 
+// Limits for input and output vector transfer (CK Tile).
+template <auto Value>
+concept TileInputOutputVectorTransferLimits =
+    requires { requires Value.a > 0 && Value.b > 0 && Value.c > 0; };
+
 // Limits for output vector transfer.
 template <auto Value>
 concept OutputVectorTransferLimits = requires {
-    requires Value.scalar_per_vector > 0 && Value.m_per_wave_per_shuffle > 0 &&
-                     Value.n_per_wave_per_shuffle > 0;
+    requires Value.scalar_per_vector > 0 && Value.m_xdl_per_wave_per_shuffle > 0 &&
+                     Value.n_xdl_per_wave_per_shuffle > 0;
 };
 
 // Limits for access order. Must be a permutation of {0, 1, 2}.